Add code for installing ZIPs directly via the app
This is compatible with all flashable Xposed ZIPs, probably even with
unofficial ones. It obviously requires root access and a modifiable
/system partition. Systemless is still WIP and might need a few more
changes in the future.

If the interface is usable, other installation methods (auto/manual
flashing) could be re-implemented in the same way. Additionally, the
checks could be extended to check compatiblility with the current ROM.